--- server/grilledshrimp/src/main/java/com/mainproject/grilledshrimp/domain/post/dto/PostPatchDto.java
@@ -1,8 +1,0 @@
-package com.mainproject.grilledshrimp.domain.post.dto;
-
-// 게시글 수정
-public class PostPatchDto {
-    private Long user_id;
-    private String post_title;
-    private String post_caption;
-}
