--- server/grilledshrimp/src/main/java/com/mainproject/grilledshrimp/domain/post/dto/PostPostDto.java
@@ -1,9 +1,0 @@
-package com.mainproject.grilledshrimp.domain.post.dto;
-
-// 게시글 생성
-public class PostPostDto {
-    private Long user_id;
-    private String post_title;
-    private String post_image;
-    private String post_caption;
-}
