+++ server/grilledshrimp/src/main/java/com/mainproject/grilledshrimp/global/image/AwsS3Service.java
@@ -1,0 +1,69 @@
+package com.mainproject.grilledshrimp.global.image;
+
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.*;
+import com.mainproject.grilledshrimp.global.exception.BusinessLogicException;
+import com.mainproject.grilledshrimp.global.exception.ExceptionCode;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.InputStreamResource;
+import org.springframework.core.io.Resource;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
+@Service
+@RequiredArgsConstructor
+public class AwsS3Service {
+
+    @Value("${cloud.aws.s3.bucket}")
+    private String bucket;
+
+    private final AmazonS3 amazonS3;
+
+    public String uploadImage(MultipartFile multipartFile) {
+        String fileName = StringUtils.cleanPath(multipartFile.getOriginalFilename());
+        // forEach 구문을 통해 multipartFile로 넘어온 파일들 하나씩 fileNameList에 추가
+
+        try {
+            ObjectMetadata objectMetadata = new ObjectMetadata();
+            objectMetadata.setContentLength(multipartFile.getSize());
+            objectMetadata.setContentType(multipartFile.getContentType());
+            objectMetadata.setContentDisposition("inline");
+            amazonS3.putObject(new PutObjectRequest(bucket, fileName, multipartFile.getInputStream(), objectMetadata)
+                    .withCannedAcl(CannedAccessControlList.PublicRead));
+
+        } catch(IOException e) {
+            throw new BusinessLogicException(ExceptionCode.IMAGE_UPLOAD_FAILED);
+        }
+
+        String fileUrl = amazonS3.getUrl(bucket, fileName).toString();
+
+        return fileUrl;
+    }
+//
+//
+//    public void deleteImage(String fileName) {
+//        amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
+//    }
+//
+//    private String createFileName(String fileName) { // 먼저 파일 업로드 시, 파일명을 난수화하기 위해 random으로 돌립니다.
+//        return UUID.randomUUID().toString().concat(getFileExtension(fileName));
+//    }
+//
+//    private String getFileExtension(String fileName) { // file 형식이 잘못된 경우를 확인하기 위해 만들어진 로직이며, 파일 타입과 상관없이 업로드할 수 있게 하기 위해 .의 존재 유무만 판단하였습니다.
+//        try {
+//            return fileName.substring(fileName.lastIndexOf("."));
+//        } catch (StringIndexOutOfBoundsException e) {
+//            throw new BusinessLogicException(ExceptionCode.ILLIGAL_IMAGE_TYPE);
+//        }
+//    }
+}
